+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Services/CustomPostTypes/CustomPostTypeInterface.php
@@ -18,4 +18,5 @@ interface CustomPostTypeInterface extends ServiceInterface
      */
     public function registerCustomPostType(): void;
     public function isHierarchical(): bool;
+    public function useCustomPostExcerptAsDescription(): bool;
 }